Move databaseFetcher.ts to /databases/database-fetcher.ts
