+++ extensions/ql-vscode/src/skeleton-query-wizard.ts
@@ -13,7 +13,10 @@ import { getErrorMessage } from "./pure/helpers-pure";
 import { QlPackGenerator } from "./qlpack-generator";
 import { DatabaseItem, DatabaseManager } from "./local-databases";
 import { ProgressCallback, UserCancellationException } from "./progress";
-import { askForGitHubRepo, downloadGitHubDatabase } from "./databaseFetcher";
+import {
+  askForGitHubRepo,
+  downloadGitHubDatabase,
+} from "./databases/database-fetcher";
 import {
   getSkeletonWizardFolder,
   isCodespacesTemplate,
