+++ extensions/ql-vscode/test/vscode-tests/cli-integration/skeleton-query-wizard.test.ts
@@ -18,7 +18,7 @@ import {
   DatabaseManager,
   FullDatabaseOptions,
 } from "../../../src/local-databases";
-import * as databaseFetcher from "../../../src/databaseFetcher";
+import * as databaseFetcher from "../../../src/databases/database-fetcher";
 import { createMockDB } from "../../factories/databases/databases";
 import { asError } from "../../../src/pure/helpers-pure";
 import { Setting } from "../../../src/config";
